Start pulling fmt out of headers
