+++ compiler+runtime/src/cpp/clojure/core_native.cpp
@@ -1,3 +1,4 @@
+#include <jank/native_persistent_string/fmt.hpp>
 #include <clojure/core_native.hpp>
 #include <jank/runtime/convert.hpp>
 #include <jank/runtime/core.hpp>
