+++ compiler+runtime/src/cpp/jank/analyze/processor.cpp
@@ -4,6 +4,7 @@
 
 #include <fmt/core.h>
 
+#include <jank/native_persistent_string/fmt.hpp>
 #include <jank/runtime/visit.hpp>
 #include <jank/runtime/context.hpp>
 #include <jank/runtime/behavior/number_like.hpp>
