+++ compiler+runtime/src/cpp/jank/codegen/llvm_processor.cpp
@@ -9,6 +9,7 @@
 #include <llvm/Transforms/Scalar/Reassociate.h>
 #include <llvm/Transforms/Scalar/SimplifyCFG.h>
 
+#include <jank/native_persistent_string/fmt.hpp>
 #include <jank/codegen/llvm_processor.hpp>
 #include <jank/runtime/context.hpp>
 #include <jank/runtime/core.hpp>
