+++ compiler+runtime/src/cpp/jank/evaluate.cpp
@@ -1,5 +1,6 @@
 #include <llvm/ExecutionEngine/Orc/LLJIT.h>
 
+#include <jank/native_persistent_string/fmt.hpp>
 #include <jank/runtime/context.hpp>
 #include <jank/runtime/ns.hpp>
 #include <jank/runtime/visit.hpp>
