+++ compiler+runtime/src/cpp/jank/runtime/core.cpp
@@ -1,3 +1,4 @@
+#include <jank/native_persistent_string/fmt.hpp>
 #include <jank/runtime/core.hpp>
 #include <jank/runtime/visit.hpp>
 #include <jank/runtime/behavior/comparable.hpp>
