+++ compiler+runtime/src/cpp/jank/runtime/core/to_string.cpp
@@ -1,4 +1,5 @@
-#include <jank/runtime/core.hpp>
+#include <jank/native_persistent_string/fmt.hpp>
+#include <jank/runtime/core/to_string.hpp>
 #include <jank/runtime/visit.hpp>
 
 namespace jank::runtime
