+++ compiler+runtime/src/cpp/jank/runtime/perf.cpp
@@ -1,5 +1,6 @@
 #include <nanobench.h>
 
+#include <jank/native_persistent_string/fmt.hpp>
 #include <jank/runtime/perf.hpp>
 #include <jank/runtime/visit.hpp>
 #include <jank/runtime/context.hpp>
