+++ compiler+runtime/src/cpp/jank/util/clang_format.cpp
@@ -1,8 +1,12 @@
+#include <fstream>
+
+#include <fmt/format.h>
+
+#include <clang/Format/Format.h>
+
 #include <jank/util/clang_format.hpp>
 #include <jank/util/process_location.hpp>
 #include <jank/util/dir.hpp>
-
-#include <clang/Format/Format.h>
 
 namespace jank::util
 {
