+++ compiler+runtime/src/cpp/jank/util/escape.cpp
@@ -1,3 +1,5 @@
+#include <fmt/format.h>
+
 #include <jank/util/escape.hpp>
 
 namespace jank::util
